+++ CHANGELOG.md
@@ -31,6 +31,7 @@ for now removed features.
   - fix autocomplete server-mode problem that was clearing the filter unintentionally
   - improve autocomplete server-mode documentation
 - **com-nav**: close mobile menu when click on a link (`wcs-com-nav-item`)
+- **native-select**: no label/description/error + required propagation from form-field
 
 ### Security